analyzejd/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
openai
sqlalchemy
psycopg2-binary
gunicorn
python-multipart
